split CLObjects into different files
