+++ pytissueoptics/rayscattering/opencl/CLObjects/scene.py
@@ -1,144 +1,7 @@
 from typing import List, NamedTuple
 
-import numpy as np
-from numpy.lib import recfunctions as rfn
-
-try:
-    import pyopencl as cl
-    import pyopencl.tools
-except ImportError:
-    pass
-
-from pytissueoptics.rayscattering.materials.scatteringMaterial import ScatteringMaterial
+from pytissueoptics.rayscattering.opencl.CLObjects.CLObject import *
 from pytissueoptics.scene.geometry import BoundingBox, Vertex, Vector
-
-
-class CLObject:
-    def __init__(self, name: str = None, struct: np.dtype = None, skipDeclaration: bool = False, buildOnce: bool = False):
-        self._name = name
-        self._struct = struct
-        self._declaration = None
-        self._dtype = None
-        self._skipDeclaration = skipDeclaration
-        self._buildOnce = buildOnce
-
-        self._HOST_buffer = None
-        self._DEVICE_buffer = None
-
-    def build(self, device: 'cl.Device', context):
-        if self.deviceBuffer is not None:
-            if self._buildOnce:
-                return
-        self.make(device)
-        self._DEVICE_buffer = cl.Buffer(context, cl.mem_flags.READ_WRITE | cl.mem_flags.USE_HOST_PTR,
-                                        hostbuf=self.hostBuffer)
-
-    def make(self, device):
-        if self._struct:
-            cl_struct, self._declaration = cl.tools.match_dtype_to_c_struct(device, self._name, self._struct)
-            self._dtype = cl.tools.get_or_register_dtype(self._name, cl_struct)
-
-    def reset(self):
-        self.hostBuffer = self._getInitialHostBuffer()
-
-    def _getInitialHostBuffer(self) -> np.ndarray:
-        raise NotImplementedError()
-
-    @property
-    def name(self) -> str:
-        return self._name
-
-    @property
-    def declaration(self) -> str:
-        if not self._declaration or self._skipDeclaration:
-            return ''
-        return self._declaration
-
-    @property
-    def dtype(self) -> ...:
-        assert self._dtype is not None
-        return self._dtype
-
-    @property
-    def hostBuffer(self):
-        if self._HOST_buffer is None:
-            self._HOST_buffer = self._getInitialHostBuffer()
-        return self._HOST_buffer
-
-    @hostBuffer.setter
-    def hostBuffer(self, value):
-        if isinstance(value, np.ndarray):
-            self._HOST_buffer = value
-        else:
-            raise TypeError("hostBuffer must be a numpy.ndarray")
-
-    @property
-    def deviceBuffer(self):
-        return self._DEVICE_buffer
-
-    @property
-    def length(self) -> int:
-        return len(self._HOST_buffer)
-
-
-class PhotonCL(CLObject):
-    STRUCT_NAME = "Photon"
-
-    def __init__(self, positions: np.ndarray, directions: np.ndarray,
-                 materialID: int, solidID: int):
-        self._positions = positions
-        self._directions = directions
-        self._N = positions.shape[0]
-        self._materialID = materialID
-        self._solidID = solidID
-
-        photonStruct = np.dtype(
-            [("position", cl.cltypes.float4),
-             ("direction", cl.cltypes.float4),
-             ("er", cl.cltypes.float4),
-             ("weight", cl.cltypes.float),
-             ("materialID", cl.cltypes.uint),
-             ("solidID", cl.cltypes.uint)])
-        super().__init__(name=self.STRUCT_NAME, struct=photonStruct)
-
-    def _getInitialHostBuffer(self) -> np.ndarray:
-        buffer = np.zeros(self._N, dtype=self._dtype)
-        buffer = rfn.structured_to_unstructured(buffer)
-        buffer[:, 0:3] = self._positions
-        buffer[:, 4:7] = self._directions
-        buffer = rfn.unstructured_to_structured(buffer, self._dtype)
-        buffer["weight"] = 1.0
-        buffer["materialID"] = self._materialID
-        buffer["solidID"] = self._solidID
-        return buffer
-
-
-class MaterialCL(CLObject):
-    STRUCT_NAME = "Material"
-
-    def __init__(self, materials: List[ScatteringMaterial]):
-        self._materials = materials
-
-        materialStruct = np.dtype(
-            [("mu_s", cl.cltypes.float),
-             ("mu_a", cl.cltypes.float),
-             ("mu_t", cl.cltypes.float),
-             ("g", cl.cltypes.float),
-             ("n", cl.cltypes.float),
-             ("albedo", cl.cltypes.float)])
-        super().__init__(name=self.STRUCT_NAME, struct=materialStruct, buildOnce=True)
-
-    def _getInitialHostBuffer(self) -> np.ndarray:
-        # todo: there might be a way to abstract both struct and buffer under a single def (DRY, PO)
-        buffer = np.empty(len(self._materials), dtype=self._dtype)
-        for i, material in enumerate(self._materials):
-            buffer[i]["mu_s"] = np.float32(material.mu_s)
-            buffer[i]["mu_a"] = np.float32(material.mu_a)
-            buffer[i]["mu_t"] = np.float32(material.mu_t)
-            buffer[i]["g"] = np.float32(material.g)
-            buffer[i]["n"] = np.float32(material.n)
-            buffer[i]["albedo"] = np.float32(material.getAlbedo())
-        return buffer
 
 
 SolidCLInfo = NamedTuple("SolidInfo", [("bbox", BoundingBox),
@@ -259,34 +122,6 @@ class VertexCL(CLObject):
                 buffer[i]["normal"][1] = np.float32(vertex.normal.y)
                 buffer[i]["normal"][2] = np.float32(vertex.normal.z)
         return buffer
-
-
-class DataPointCL(CLObject):
-    STRUCT_NAME = "DataPoint"
-
-    def __init__(self, size: int):
-        self._size = size
-
-        dataPointStruct = np.dtype(
-            [("delta_weight", cl.cltypes.float),
-             ("x", cl.cltypes.float),
-             ("y", cl.cltypes.float),
-             ("z", cl.cltypes.float),
-             ("solidID", cl.cltypes.uint),
-             ("surfaceID", cl.cltypes.int)])
-        super().__init__(name=self.STRUCT_NAME, struct=dataPointStruct)
-
-    def _getInitialHostBuffer(self) -> np.ndarray:
-        return np.zeros(self._size, dtype=self._dtype)
-
-
-class SeedCL(CLObject):
-    def __init__(self, size: int):
-        self._size = size
-        super().__init__(buildOnce=True)
-
-    def _getInitialHostBuffer(self) -> np.ndarray:
-        return np.random.randint(low=0, high=2 ** 32 - 1, size=self._size, dtype=cl.cltypes.uint)
 
 
 class SolidCandidateCL(CLObject):
